1. Cannot buy until no money 2. Bomb does damages now instead of insta-kill 3. Other turrets have sound 4. Particles will despawn(not un-render) when in the air for too long 5. No more reference axis nor lightball

+++ StudioProject3/Base/Source/GameUI/WinLoseScreen.cpp
@@ -141,7 +141,7 @@ void WinLoseScreen::Update(double dt)
 
 	if (GetSwitchLevel())
 	{
-		m_fTimeBeforeLevelChange -= dt;
+		m_fTimeBeforeLevelChange -= (float)dt;
 
 		std::ostringstream ss;
 		ss.precision(1);
